refactor(BlockRenderer): move BlockRenderer to PageBuilder directory and update imports
- Relocated BlockRenderer component to a new PageBuilder directory for better organization.
- Updated import paths in Home and Page components to reflect the new location of BlockRenderer.
- Introduced a new BlockRenderer implementation that supports various block types for dynamic rendering.
